pyproject.toml
[project]
name = "city-renewal-open-source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "kafka>=1.3.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "socketio>=0.2.1",
]
